Revert "Merge pull request #56602 from Amoralchik/clickable-emoji-for-exoense-task"
This reverts commit a1fae232fab21b6fbee3db512119fdac1b953c7d, reversing
changes made to d47c197666aca89da31b512e8bc350b88dfe8b9b.
