+++ src/libs/Navigation/NavigationRoot.tsx
@@ -3,7 +3,6 @@ import {DarkTheme, DefaultTheme, findFocusedRoute, NavigationContainer} from '@r
 import React, {useContext, useEffect, useMemo, useRef} from 'react';
 import {NativeModules} from 'react-native';
 import {useOnyx} from 'react-native-onyx';
-import FABPopoverProvider from '@components/FABPopoverProvider';
 import {ScrollOffsetContext} from '@components/ScrollOffsetContextProvider';
 import {usePlaybackContext} from '@components/VideoPlayerContexts/PlaybackContext';
 import useCurrentReportID from '@hooks/useCurrentReportID';
@@ -224,9 +223,7 @@ function NavigationRoot({authenticated, lastVisitedPath, initialUrl, onReady, sh
                 enabled: false,
             }}
         >
-            <FABPopoverProvider>
-                <AppNavigator authenticated={authenticated} />
-            </FABPopoverProvider>
+            <AppNavigator authenticated={authenticated} />
         </NavigationContainer>
     );
 }
