+++ src/pages/home/sidebar/BottomTabBarFloatingActionButton/index.native.tsx
@@ -1,0 +1,3 @@
+import FloatingActionButtonAndPopover from '@pages/home/sidebar/FloatingActionButtonAndPopover';
+
+export default FloatingActionButtonAndPopover;
