+++ src/pages/home/sidebar/FloatingActionButtonAndPopover.tsx
@@ -1,14 +1,14 @@
 import {useIsFocused} from '@react-navigation/native';
 import type {ImageContentFit} from 'expo-image';
 import type {ForwardedRef} from 'react';
-import React, {forwardRef, useCallback, useContext, useEffect, useImperativeHandle, useMemo, useState} from 'react';
-import {NativeModules} from 'react-native';
+import React, {forwardRef, useCallback, useContext, useEffect, useImperativeHandle, useMemo, useRef, useState} from 'react';
+import {NativeModules, View} from 'react-native';
 import type {OnyxCollection, OnyxEntry} from 'react-native-onyx';
 import {useOnyx} from 'react-native-onyx';
 import type {SvgProps} from 'react-native-svg';
 import ConfirmModal from '@components/ConfirmModal';
 import CustomStatusBarAndBackgroundContext from '@components/CustomStatusBarAndBackground/CustomStatusBarAndBackgroundContext';
-import {FABPopoverContext} from '@components/FABPopoverProvider';
+import FloatingActionButton from '@components/FloatingActionButton';
 import * as Expensicons from '@components/Icon/Expensicons';
 import type {PopoverMenuItem} from '@components/PopoverMenu';
 import PopoverMenu from '@components/PopoverMenu';
@@ -51,7 +51,18 @@ import mapOnyxCollectionItems from '@src/utils/mapOnyxCollectionItems';
 
 type PolicySelector = Pick<OnyxTypes.Policy, 'type' | 'role' | 'isPolicyExpenseChatEnabled' | 'pendingAction' | 'avatarURL' | 'name' | 'id' | 'areInvoicesEnabled'>;
 
-type FloatingActionButtonPopoverRef = {
+type FloatingActionButtonAndPopoverProps = {
+    /* Callback function when the menu is shown */
+    onShowCreateMenu?: () => void;
+
+    /* Callback function before the menu is hidden */
+    onHideCreateMenu?: () => void;
+
+    /* If the tooltip is allowed to be shown */
+    isTooltipAllowed: boolean;
+};
+
+type FloatingActionButtonAndPopoverRef = {
     hideCreateMenu: () => void;
 };
 
@@ -150,7 +161,7 @@ const getQuickActionTitle = (action: QuickActionName): TranslationPaths => {
  * Responsible for rendering the {@link PopoverMenu}, and the accompanying
  * FAB that can open or close the menu.
  */
-function FloatingActionButtonPopover(_: unknown, ref: ForwardedRef<FloatingActionButtonPopoverRef>) {
+function FloatingActionButtonAndPopover({onHideCreateMenu, onShowCreateMenu, isTooltipAllowed}: FloatingActionButtonAndPopoverProps, ref: ForwardedRef<FloatingActionButtonAndPopoverRef>) {
     const styles = useThemeStyles();
     const theme = useTheme();
     const {translate} = useLocalize();
@@ -173,8 +184,9 @@ function FloatingActionButtonPopover(_: unknown, ref: ForwardedRef<FloatingActio
     const [quickActionPolicy] = useOnyx(`${ONYXKEYS.COLLECTION.POLICY}${quickActionReport?.policyID}`);
     const [allPolicies] = useOnyx(ONYXKEYS.COLLECTION.POLICY, {selector: (c) => mapOnyxCollectionItems(c, policySelector)});
 
-    const {isCreateMenuActive, hideCreateMenu, fabRef} = useContext(FABPopoverContext);
+    const [isCreateMenuActive, setIsCreateMenuActive] = useState(false);
     const [modalVisible, setModalVisible] = useState(false);
+    const fabRef = useRef<HTMLDivElement>(null);
     const {windowHeight} = useWindowDimensions();
     const {shouldUseNarrowLayout} = useResponsiveLayout();
     const isFocused = useIsFocused();
@@ -268,6 +280,38 @@ function FloatingActionButtonPopover(_: unknown, ref: ForwardedRef<FloatingActio
         [isFocused, prevIsFocused],
     );
 
+    /**
+     * Method called when we click the floating action button
+     */
+    const showCreateMenu = useCallback(
+        () => {
+            if (!isFocused && shouldUseNarrowLayout) {
+                return;
+            }
+            setIsCreateMenuActive(true);
+            onShowCreateMenu?.();
+        },
+        // eslint-disable-next-line react-compiler/react-compiler, react-hooks/exhaustive-deps
+        [isFocused, shouldUseNarrowLayout],
+    );
+
+    /**
+     * Method called either when:
+     * - Pressing the floating action button to open the CreateMenu modal
+     * - Selecting an item on CreateMenu or closing it by clicking outside of the modal component
+     */
+    const hideCreateMenu = useCallback(
+        () => {
+            if (!isCreateMenuActive) {
+                return;
+            }
+            setIsCreateMenuActive(false);
+            onHideCreateMenu?.();
+        },
+        // eslint-disable-next-line react-compiler/react-compiler, react-hooks/exhaustive-deps
+        [isCreateMenuActive],
+    );
+
     useEffect(() => {
         if (!didScreenBecomeInactive()) {
             return;
@@ -282,6 +326,14 @@ function FloatingActionButtonPopover(_: unknown, ref: ForwardedRef<FloatingActio
             hideCreateMenu();
         },
     }));
+
+    const toggleCreateMenu = () => {
+        if (isCreateMenuActive) {
+            hideCreateMenu();
+        } else {
+            showCreateMenu();
+        }
+    };
 
     const expenseMenuItems = useMemo((): PopoverMenuItem[] => {
         return [
@@ -489,10 +541,10 @@ function FloatingActionButtonPopover(_: unknown, ref: ForwardedRef<FloatingActio
     ];
 
     return (
-        <>
+        <View style={styles.flexGrow1}>
             <PopoverMenu
                 onClose={hideCreateMenu}
-                isVisible={isCreateMenuActive}
+                isVisible={isCreateMenuActive && (!shouldUseNarrowLayout || isFocused)}
                 anchorPosition={styles.createMenuPositionSidebar(windowHeight)}
                 onItemSelected={hideCreateMenu}
                 fromSidebarMediumScreen={!shouldUseNarrowLayout}
@@ -530,12 +582,20 @@ function FloatingActionButtonPopover(_: unknown, ref: ForwardedRef<FloatingActio
                 confirmText={translate('exitSurvey.goToExpensifyClassic')}
                 cancelText={translate('common.cancel')}
             />
-        </>
+            <FloatingActionButton
+                isTooltipAllowed={isTooltipAllowed}
+                accessibilityLabel={translate('sidebarScreen.fabNewChatExplained')}
+                role={CONST.ROLE.BUTTON}
+                isActive={isCreateMenuActive}
+                ref={fabRef}
+                onPress={toggleCreateMenu}
+            />
+        </View>
     );
 }
 
-FloatingActionButtonPopover.displayName = 'FloatingActionButtonPopover';
+FloatingActionButtonAndPopover.displayName = 'FloatingActionButtonAndPopover';
 
-export default forwardRef(FloatingActionButtonPopover);
+export default forwardRef(FloatingActionButtonAndPopover);
 
 export type {PolicySelector};
