+++ src/styles/index.ts
@@ -13,7 +13,6 @@ import type {ValueOf} from 'type-fest';
 import type DotLottieAnimation from '@components/LottieAnimations/types';
 import {ACTIVE_LABEL_SCALE} from '@components/TextInput/styleConst';
 import {getBrowser, isMobile, isMobileSafari, isSafari} from '@libs/Browser';
-import getPlatform from '@libs/getPlatform';
 import CONST from '@src/CONST';
 import {defaultTheme} from './theme';
 import colors from './theme/colors';
@@ -372,9 +371,6 @@ const styles = (theme: ThemeColors) =>
         },
         verticalAlignTop: {
             verticalAlign: 'top',
-        },
-        verticalAlignBottom: {
-            verticalAlign: 'bottom',
         },
         lineHeightLarge: {
             lineHeight: variables.lineHeightLarge,
@@ -1705,16 +1701,6 @@ const styles = (theme: ThemeColors) =>
             borderRadius: 999,
             alignItems: 'center',
             justifyContent: 'center',
-        },
-
-        customEmoji: {
-            alignItems: 'center',
-            justifyContent: 'center',
-            verticalAlign: 'bottom',
-            ...(getPlatform() === CONST.PLATFORM.IOS ? {marginBottom: -variables.fontSizeNormalHeight / 4} : {}),
-            // `fontSizeNormal` value for scale factor 1 is 15
-            ...(getPlatform() === CONST.PLATFORM.ANDROID ? {marginBottom: -(variables.fontSizeEmojiNormal - 10)} : {}),
-            ...(getPlatform() === CONST.PLATFORM.ANDROID ? {marginRight: -(variables.fontSizeEmojiNormal - 15)} : {}),
         },
 
         sidebarFooterUsername: {
